// src/app/ui/todoConsola/components/ContentTitleMenu.enum.ts
export enum ContentTitleMenu{
  titleCreateModulo = 'Por favor ingrese un modulo...',
  titleCreateTask = 'Inserte la tarea a realizar...',
  titleRead = '---- Read Task ----',
  titleUpdate = '--- Update Task ---',
  titleUpdatePosition = 'escoja la posicion de la tarea que desea actualizar',
  titleUpdateModulo = 'Inserte el modulo referente a la tarea para actualizar',
  titleUpdateTask = 'Inserte la tarea referente para actualizar',
  titleDeletePosition = 'escoja la posicion de la tarea a eliminar',
  titleExitApp = 'thank you for using our TODO'
}